Add models Add services Add repositories Add controllers

+++ src/main/java/com/davorin/onboarding/repository/ProcessRepository.java
@@ -12,5 +12,5 @@ public interface ProcessRepository {
 
     Long getSequence();
 
-    Process getProcessByUser(int userId);
+    Process getProcessByUser(Long userId);
 }
